Add helpers to Fatalities component, move more query URLs to query file
