+++ atd-vzv/src/constants/responsive.js
@@ -1,4 +1,4 @@
-// Use sizes below to convert Material-UI breakpoints to pixels
+// Use to convert Material-UI breakpoints to pixels
 export const responsive = {
   sm: 600,
   md: 960
