+++ atd-vzv/src/views/dashboard/fatalities.js
@@ -1,6 +1,11 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
-import moment from "moment";
+import { calculateTotalFatalities } from "./helpers/helpers";
+import { thisYear, lastYear } from "./helpers/time";
+import {
+  fatalitiesYTDUrl,
+  previousYearFatalitiesUrl
+} from "./queries/socrataQueries";
 
 import { Container, Row, Col } from "reactstrap";
 import styled from "styled-components";
@@ -12,27 +17,11 @@ const StyledIcon = styled.i`
 `;
 
 const Fatalities = () => {
-  const today = moment().format("YYYY-MM-DD");
-  const todayMonthYear = moment().format("-MM-DD");
-  const thisYear = moment().format("YYYY");
-  const lastYear = moment()
-    .subtract(1, "year")
-    .format("YYYY");
-
-  const yearToDateUrl = `https://data.austintexas.gov/resource/y2wy-tgr5.json?$where=death_cnt > 0 AND crash_date between '${thisYear}-01-01T00:00:00' and '${today}T23:59:59'`;
-  const previousYearUrl = `https://data.austintexas.gov/resource/y2wy-tgr5.json?$where=death_cnt > 0 AND crash_date between '${lastYear}-01-01T00:00:00' and '${lastYear}${todayMonthYear}T23:59:59'`;
-
   const [yearToDateFatalityTotal, setYearToDateFatalityTotal] = useState(0);
   const [
     lastYearToDateFatalityTotal,
     setLastYearToDateFatalityTotal
   ] = useState(0);
-
-  const calculateTotalFatalities = data => {
-    let total = 0;
-    data.data.forEach(record => (total += parseInt(record.death_cnt)));
-    return total;
-  };
 
   const renderFatalityIcons = fatalityTotal =>
     // Create array with length of fatalityTotal and use to iterate
@@ -44,15 +33,15 @@ const Fatalities = () => {
 
   useEffect(() => {
     // Fetch year-to-date records
-    axios.get(yearToDateUrl).then(res => {
-      setYearToDateFatalityTotal(calculateTotalFatalities(res));
+    axios.get(fatalitiesYTDUrl).then(res => {
+      setYearToDateFatalityTotal(calculateTotalFatalities(res.data));
     });
 
     // Fetch last year-to-date records
-    axios.get(previousYearUrl).then(res => {
-      setLastYearToDateFatalityTotal(calculateTotalFatalities(res));
+    axios.get(previousYearFatalitiesUrl).then(res => {
+      setLastYearToDateFatalityTotal(calculateTotalFatalities(res.data));
     });
-  }, [yearToDateUrl, previousYearUrl]);
+  }, []);
 
   return (
     <Container>
