+++ atd-vzv/src/views/dashboard/helpers/time.js
@@ -1,5 +1,9 @@
 import moment from "moment";
 
 // Time data
-export const thisYear = moment().format("YYYY");
 export const today = moment().format("YYYY-MM-DD");
+export const todayMonthYear = moment().format("-MM-DD");
+export const thisYear = moment().format("YYYY");
+export const lastYear = moment()
+  .subtract(1, "year")
+  .format("YYYY");
